change files to let error persist

+++ src/chapter21miscconcepts/CountNumberofSetbitsin1toN.java
@@ -11,19 +11,18 @@
  * 
  */
 
-
 package chapter21miscconcepts;
 
 public class CountNumberofSetbitsin1toN {
 	public int countNumberofSetbitsin1toN(int n) {
-		long i = 0, j, count = 0; 
-		for (i = 1; i<=n; i++){
+		long i = 0, j, count = 0;
+		for (i = 1; i <= n; i++) {
 			j = i;
-			while(j>0) {
-			     count++;
-			     j = j & (j - 1); 
+			while (j > 0) {
+				count++;
+				j = j & (j - 1);
 			}
 		}
-		return count;		
+		return (int) count;
 	}
 }
